Service image uses project target_registry
